views : add check for superusers in layer update notifications
When a user submits a layer, notification emails are sent to users
with a certain permission. Superusers technically have all
permissions, but Q ignores them currently so that such emails
are not sent to superusers. This patch expands the check to
include superusers.

A few folks have discussed including this so I put this patch
up for consideration.